+++ src/__tests__/send-request.js
@@ -1,10 +1,12 @@
-import { encodePostData, xhr } from '../send-request'
+import { addParamsToURL, encodePostData, xhr } from '../send-request'
 import { assert, boolean, property, uint8Array, VerbosityLevel } from 'fast-check'
 
-describe('when xhr requests fail', () => {
+jest.mock('../config', () => ({ DEBUG: false, LIB_VERSION: '1.23.45' }))
+
+describe('xhr', () => {
     given('mockXHR', () => ({
         open: jest.fn(),
-        setRequestHeader: jest.fn(),
+        setRequestHeader: jest.fn,
         onreadystatechange: jest.fn(),
         send: jest.fn(),
         readyState: 4,
@@ -37,17 +39,43 @@ describe('when xhr requests fail', () => {
         window.XMLHttpRequest = jest.fn(() => given.mockXHR)
     })
 
-    it('does not error if the configured onXHRError is not a function', () => {
-        given('onXHRError', () => 'not a function')
-        expect(() => given.subject()).not.toThrow()
+    describe('when xhr requests fail', () => {
+        it('does not error if the configured onXHRError is not a function', () => {
+            given('onXHRError', () => 'not a function')
+            expect(() => given.subject()).not.toThrow()
+        })
+
+        it('calls the injected XHR error handler', () => {
+            //cannot use an auto-mock from jest as the code checks if onXHRError is a Function
+            let requestFromError
+            given('onXHRError', () => (req) => (requestFromError = req))
+            given.subject()
+            expect(requestFromError).toHaveProperty('status', 502)
+        })
+    })
+})
+
+describe('adding query params to posthog API calls', () => {
+    given('subject', () => () => addParamsToURL(given.posthogURL, given.urlQueryArgs, given.parameterOptions))
+    given('posthogURL', () => 'https://any.posthog-instance.com')
+    given('urlQueryArgs', () => ({}))
+    given('parameterOptions', () => ({
+        ip: true,
+    }))
+
+    it('adds library and version', () => {
+        expect(new URL(given.subject()).search).toContain('&v=1.23.45')
     })
 
-    it('calls the injected XHR error handler', () => {
-        //cannot use an auto-mock from jest as the code checks if onXHRError is a Function
-        let requestFromError
-        given('onXHRError', () => (req) => (requestFromError = req))
-        given.subject()
-        expect(requestFromError).toHaveProperty('status', 502)
+    it('adds i as 1 when IP in config', () => {
+        expect(new URL(given.subject()).search).toContain('ip=1')
+    })
+    it('adds i as 0 when IP not in config', () => {
+        given('parameterOptions', () => ({}))
+        expect(new URL(given.subject()).search).toContain('ip=0')
+    })
+    it('adds timestamp', () => {
+        expect(new URL(given.subject()).search).toMatch(/_=\d+/)
     })
 })