+++ src/extensions/sessionrecording.js
@@ -2,13 +2,14 @@ import { loadScript } from '../autocapture-utils'
 import { _ } from '../utils'
 import { SESSION_RECORDING_ENABLED_SERVER_SIDE } from '../posthog-persistence'
 import Config from '../config'
-import { filterDataURLsFromLargeDataObjects } from './sessionrecording-utils'
+import { filterDataURLsFromLargeDataObjects, truncateLargeConsoleLogs } from './sessionrecording-utils'
 
 const BASE_ENDPOINT = '/e/'
 
 export const FULL_SNAPSHOT_EVENT_TYPE = 2
 export const META_EVENT_TYPE = 4
 export const INCREMENTAL_SNAPSHOT_EVENT_TYPE = 3
+export const PLUGIN_EVENT_TYPE = 6
 export const MUTATION_SOURCE_TYPE = 0
 
 export class SessionRecording {
@@ -145,7 +146,7 @@ export class SessionRecording {
 
         this.stopRrweb = this.rrwebRecord({
             emit: (event) => {
-                event = filterDataURLsFromLargeDataObjects(event)
+                event = truncateLargeConsoleLogs(filterDataURLsFromLargeDataObjects(event))
 
                 this._updateWindowAndSessionIds(event)